llamaindexqa-develop/requirements.txt
numpy~=1.24.4
pandas~=1.5
streamlit>=1.25.0
streamlit-option-menu>=0.3.6
streamlit-antd-components>=0.1.11
streamlit-chatbox>=1.1.6
transformers==4.31.0
pypdf2
nvitop==1.2.0
langchain
faiss-cpu==1.7.2
openai
parsr-client
streamlit_extras
streamlit_javascript
streamlit-echarts
stqdm
matplotlib
scikit-learn
wordcloud
python-dotenv
loguru
omegaconf==2.1.1
PyMuPDF
hanlp[full]